Extract BubbleChart titles
